App folder renamed
